requirements.txt
Django==5.0.6
psycopg2-binary==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.2
django-filter==24.3
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.6.1
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
daphne==4.1.2
channels==4.1.0
Pillow==10.4.0
numpy==1.26.4
scikit-learn==1.4.2
whitenoise==6.7.0
